[feat/#102] dto, model, mapper 추가

+++ app/src/main/java/com/smashing/app/data/model/search/SearchMainItemModel.kt
@@ -12,5 +12,5 @@ data class SearchMainItemModel(
     val tierId: TierType,
     val wins: Int,
     val losses: Int,
-    val reviews: Int,
+    val reviews: Long,
 )
